build: generate env file with component versions
When splitting up gitlab into components, we need to make sure that we
get the correct versions of all the components. These versions are
defined in the gitlab source as separate files.

We use the gitlab API to get the contents of the file and write an env
file with the version of each component. This can be made available as
build arguments to the Dockerfiles.

This uses go-task, which is a simple and straightforward way to define
common tasks.