docker/gitaly: add stage for gitaly
This builds the go components of gitaly. To prevent the make file from
also building the ruby component, the Makefile is patched to remove the
dependency. The ruby component will be built in a separate stage.

To reduce the size of the files that end up in the final image, we strip
the binaries of debug symbols using `strip`.