docker/gitaly: build final image
This copies the build artifacts from earlier images to compose the final
image. To make sure we have all soname dependencies, use scanelf to find
the sonames needed for each binary so that they can be installed.

This is easier than manually tracking dependencies and is based on the
gemdeps.so script that is used in the gitlab container.