gitaly: split up build script
With a single script that is copied in each stage, any changes forces
all build stages to be rebuilt.

Instead, split up the script in separate scripts that each git copied
just to their respective stages. The same counts for the patches.

Some functionality is shared by different build stages. We extract this
functionality in a separate lib file that can be sourced between stages
and even in different components.