docker: move all sockets to a single volume
Instead of having different volumes for different sockets, move all of
them to just a single volume.

To do this, we first make sure we can have a single mountpoint for all
sockets, which is `/home/git/run`. In there, there is a directory per
service where the socket will live. Secondly, we need to adjust the
socket locations to there.

For gitlab, we need to adjust the socket location for both
puma (provider) as workhorse (consumer), which is done through patches
to the default configuration.

Gitaly was already setup this way, so for gitaly we do not need to
change anything.