gitaly: use v3.13 alpine golang image
Although go is generally statically built, it still links against musl.
musl on Alpine Edge added support for qsort_r, and apparently golang
uses it if present.

When running these binaries on Alpine 3.13, which we still do for
gitlab, we get linking errors:

> Error relocating /usr/local/bin/gitaly-git2go-v14: qsort_r: symbol not
> found

Use golang from Alpine 3.13 to avoid this.