gitlab/entrypoint: add dedicated config command
When running a container without starting gitlab through the entrypoint,
the configuration is missing, so it's not possible to run things like
rake commands.

This might be required in case there is some issue preventing gitlab to
start, causing a restart loop.

By adding a dedicated config command, one can run the container with a
shell, run the config entrypoint command and then run something like a
rake task.